Move the request and serve tests into their own file
Everything on the serving side of a SYNC_REQUEST: the handler, the combat
door in front of it, the fingerprint diff that decides what to offer, the
bounded manifest both sides speak, send order, and the session cap that
stops a backfill running for hours.

The combat serve gate sits here rather than with the combat teardown tests
because what it pins is HandleSyncRequest's door ordering.
